Add ModDLL skeleton.
This is the first ModDLL written outside of Cyan that adds an external
creatable, as far as I am aware. The goal here is to be compatible with
both TPotS and MOUL in the same codebase. This will be tedious, but very
rewarding once we're done. Right now, we really just register the
creatable, write a log message, then leak memory on exit. Next will be
to create the plugin's key and start to receive messages from Plasma.

Note that building for TPotS is currently untested. Also, a change to
`hsRefCount`'s vftable in H-uru/Plasma needs to be backed out.